refactor program cs for readability and extensibility
